Module1/python-app/requirements.txt
flask
mysql-connector-python
boto3